+++ lib/checkbox_grouped.dart
@@ -3,8 +3,9 @@ library checkbox_grouped;
 export 'src/controller/custom_group_controller.dart';
 export 'src/controller/group_controller.dart';
 export 'src/controller/list_group_controller.dart';
+export 'src/controller/list_custom_group_controller.dart';
 export 'src/widgets/custom_grouped_checkbox.dart'
-    hide CustomGroupedCheckboxState;
+    hide CustomGroupedCheckboxState, ItemWidget;
 export 'src/widgets/list_grouped_checkbox.dart' hide ListGroupedCheckboxState;
 export 'src/widgets/show_dialog_grouped_checkbox.dart';
 export 'src/widgets/simple_grouped_checkbox.dart'
@@ -13,3 +14,4 @@ export 'src/widgets/simple_grouped_chips.dart' hide SimpleGroupedChipsState;
 export 'src/widgets/simple_grouped_switch.dart' hide SimpleGroupedSwitchState;
 export 'src/common/utilities.dart';
 export 'src/common/grouped_style.dart';
+export 'src/widgets/list_custom_grouped_checkbox.dart';